txm: postpone garbage collector
Before this patch garabage collector was executed right before
allocation of a new story. That means that, for example, in the
memtx_tx_history_add_stmt GC could be called a couple of times.

Garbage collector is free to delete stories if they are no more
used. Removing a story can cause an index modification with
further tuple delete.

For example imagine a space with one index, where one tuple
{1, 1, 1} is placed. Then a transaction comes, deletes that tuple
and commits. In this moment the tuple {1, 1, 1} can be still in
index, marked as 'dirty' and having a corresponding story, which
states that the tuple is deleted.

This a valid situation, even necessary, for the case when another
transaction is in a read view and must see that {1, 1, 1} not yet
deleted. But when possible, GC would try to delete the story and
remove the tuple from index.

Now imagine that this GC happens when a new transaction inserts,
for example, {1, 1, 1, 4}. In memtx_tx_history_add_stmt the new
tuple replaces the old one in index, but the story of new tuple
is not created yet. Then the new story is created, that causes
GC, that tries to remove {1, 1, 1} from index and delete it from
memory. An this moment memtx_tx_history_add_stmt relies on
existance of {1, 1, 1} which doesn't exist.

That is an example of general problem: a cleanup should not be
done in the middle of complex function that can have some half
made not valid intermediate state. The cleanup, including GC,
should be done in the end of functions.

This patch move story GC to the end of functions that use it.

Part of #5515