txm: fix memtx_tx_tuple_clarify_slow call args
At some point memtx_tx_tuple_clarify_slow was changed in terms of
its signarure - index ID was replaced with pointer to index.
Unfortunately one of calls of this function was not fixed - zero
was passed as index ID which compiles successfully with the new
API.

This patch fixes that place as well and adds a bunch of tests.

Closes #6229