iproto: fix crash if box.cfg listen is woken up.
There was access to previously freed memory in case when `cbus_call`
is interrupted: `cbus_call_msg` in iproto allocates on stack, and if
`cbus_call` failed due to fiber cancelation or wake up, `cbus_call_msg`
memory is released. But function called through cbus is still work in
iproto thread and there will be an attempt to access this memory when
this function in iproto thread finished it's work. This pacth rework
`cbus_call` behaviour, if TIMEOUT_INFINITY passed to `cbus_call` it
can't be woken up. Also before `cbus_call` we reset FIBER_IS_CANCELLABLE
flag, to prevent fiber cancellation.

Closes #6480